chore: merge light DID utils into light DID
